backend/requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow-cpu==2.14.0
python-multipart
